rename NavBar to Header
